+++ scripts/verify/mean-aggregation-strategy.js
@@ -1,11 +1,13 @@
 const hre = require("hardhat");
 
-const contractAddress = "0xc2861530884607C5d2D327106CA00f252f934962";
+const contractAddress = "0x7094A70883DC4Ac71650D2A84378d28F04d6De0b";
 
-const averagingStrategy = "0xd850F64Eda6a62d625209711510f43cD49Ef8798";
+const averagingStrategy = "0x8302AE5d7603E36f1078066C2e436C1ca0094610";
 
 async function main() {
     await hre.run("verify:verify", {
+        contract:
+            "contracts/strategies/aggregation/QuoteTokenWeightedMeanAggregator.sol:QuoteTokenWeightedMeanAggregator",
         address: contractAddress,
         constructorArguments: [averagingStrategy],
     });